--- a/packages/engine/src/common/proxies/createThreejsProxy.ts
+++ b/packages/engine/src/common/proxies/createThreejsProxy.ts
@@ -1,4 +1,5 @@
 import { Quaternion, Vector3 } from 'three'
+import { number } from 'ts-matches'
 
 import { Entity } from '../../ecs/classes/Entity'
 
@@ -7,38 +8,47 @@ const { defineProperties } = Object
 type Vector3Store = { x: Float64Array; y: Float64Array; z: Float64Array }
 type QuaternionStore = { x: Float64Array; y: Float64Array; z: Float64Array; w: Float64Array }
 
-export const proxifyVector3 = (store: Vector3Store, entity: Entity, vector3 = new Vector3()): Vector3 => {
+interface ProxyExtensions {
+  entity: number
+  store: Vector3Store | QuaternionStore
+}
+
+export const proxifyVector3 = (
+  store: Vector3Store,
+  entity: Entity,
+  vector3 = new Vector3()
+): Vector3 & ProxyExtensions => {
   // Set the initial values
   store.x[entity] = vector3.x
   store.y[entity] = vector3.y
   store.z[entity] = vector3.z
-  return defineProperties(vector3, {
-    _eid: { value: entity, configurable: true },
-    _store: { value: store, configurable: true },
+  return defineProperties(vector3 as Vector3 & ProxyExtensions, {
+    entity: { value: entity, configurable: true, writable: true },
+    store: { value: store, configurable: true, writable: true },
     x: {
       get() {
-        return this._store.x[this._eid]
+        return this.store.x[this.entity]
       },
       set(n) {
-        return (this._store.x[this._eid] = n)
+        return (this.store.x[this.entity] = n)
       },
       configurable: true
     },
     y: {
       get() {
-        return this._store.y[this._eid]
+        return this.store.y[this.entity]
       },
       set(n) {
-        return (this._store.y[this._eid] = n)
+        return (this.store.y[this.entity] = n)
       },
       configurable: true
     },
     z: {
       get() {
-        return this._store.z[this._eid]
+        return this.store.z[this.entity]
       },
       set(n) {
-        return (this._store.z[this._eid] = n)
+        return (this.store.z[this.entity] = n)
       },
       configurable: true
     }
@@ -48,44 +58,44 @@ export const proxifyVector3 = (store: Vector3Store, entity: Entity, vector3 = ne
 export const proxifyVector3WithDirty = (
   store: Vector3Store,
   entity: Entity,
-  dirty: Set<Entity>,
+  dirty: Record<Entity, true>,
   vector3 = new Vector3()
-): Vector3 => {
+): Vector3 & ProxyExtensions => {
   // Set the initial values
   store.x[entity] = vector3.x
   store.y[entity] = vector3.y
   store.z[entity] = vector3.z
-  dirty.add(entity)
-  return defineProperties(vector3, {
-    _eid: { value: entity, configurable: true },
-    _store: { value: store, configurable: true },
+  dirty[entity] = true
+  return defineProperties(vector3 as Vector3 & ProxyExtensions, {
+    entity: { value: entity, configurable: true, writable: true },
+    store: { value: store, configurable: true, writable: true },
     x: {
       get() {
-        return this._store.x[this._eid]
+        return this.store.x[this.entity]
       },
       set(n) {
-        dirty.add(this._eid)
-        return (this._store.x[this._eid] = n)
+        dirty[this.entity] = true
+        return (this.store.x[this.entity] = n)
       },
       configurable: true
     },
     y: {
       get() {
-        return this._store.y[this._eid]
+        return this.store.y[this.entity]
       },
       set(n) {
-        dirty.add(this._eid)
-        return (this._store.y[this._eid] = n)
+        dirty[this.entity] = true
+        return (this.store.y[this.entity] = n)
       },
       configurable: true
     },
     z: {
       get() {
-        return this._store.z[this._eid]
+        return this.store.z[this.entity]
       },
       set(n) {
-        dirty.add(this._eid)
-        return (this._store.z[this._eid] = n)
+        dirty[this.entity] = true
+        return (this.store.z[this.entity] = n)
       },
       configurable: true
     }
@@ -96,48 +106,48 @@ export const proxifyQuaternion = (
   store: QuaternionStore,
   entity: Entity,
   quaternion = new Quaternion()
-): Quaternion => {
+): Quaternion & ProxyExtensions => {
   // Set the initial values
   store.x[entity] = quaternion.x
   store.y[entity] = quaternion.y
   store.z[entity] = quaternion.z
   store.w[entity] = quaternion.w
-  return defineProperties(quaternion, {
-    _eid: { value: entity, configurable: true },
-    _store: { value: store, configurable: true },
+  return defineProperties(quaternion as Quaternion & ProxyExtensions, {
+    entity: { value: entity, configurable: true, writable: true },
+    store: { value: store, configurable: true, writable: true },
     _x: {
       get() {
-        return this._store.x[this._eid]
+        return this.store.x[this.entity]
       },
       set(n) {
-        return (this._store.x[this._eid] = n)
+        return (this.store.x[this.entity] = n)
       },
       configurable: true
     },
     _y: {
       get() {
-        return this._store.y[this._eid]
+        return this.store.y[this.entity]
       },
       set(n) {
-        return (this._store.y[this._eid] = n)
+        return (this.store.y[this.entity] = n)
       },
       configurable: true
     },
     _z: {
       get() {
-        return this._store.z[this._eid]
+        return this.store.z[this.entity]
       },
       set(n) {
-        return (this._store.z[this._eid] = n)
+        return (this.store.z[this.entity] = n)
       },
       configurable: true
     },
     _w: {
       get() {
-        return this._store.w[this._eid]
+        return this.store.w[this.entity]
       },
       set(n) {
-        return (this._store.w[this._eid] = n)
+        return (this.store.w[this.entity] = n)
       },
       configurable: true
     }
@@ -147,55 +157,55 @@ export const proxifyQuaternion = (
 export const proxifyQuaternionWithDirty = (
   store: QuaternionStore,
   entity: Entity,
-  dirty: Set<Entity>,
+  dirty: Record<Entity, true>,
   quaternion = new Quaternion()
-): Quaternion => {
+): Quaternion & ProxyExtensions => {
   // Set the initial values
   store.x[entity] = quaternion.x
   store.y[entity] = quaternion.y
   store.z[entity] = quaternion.z
   store.w[entity] = quaternion.w
-  dirty.add(entity)
-  return defineProperties(quaternion, {
-    _eid: { value: entity, configurable: true },
-    _store: { value: store, configurable: true },
+  dirty[entity] = true
+  return defineProperties(quaternion as Quaternion & ProxyExtensions, {
+    entity: { value: entity, configurable: true, writable: true },
+    store: { value: store, configurable: true, writable: true },
     _x: {
       get() {
-        return this._store.x[this._eid]
+        return this.store.x[this.entity]
       },
       set(n) {
-        dirty.add(this._eid)
-        return (this._store.x[this._eid] = n)
+        dirty[this.entity] = true
+        return (this.store.x[this.entity] = n)
       },
       configurable: true
     },
     _y: {
       get() {
-        return this._store.y[this._eid]
+        return this.store.y[this.entity]
       },
       set(n) {
-        dirty.add(this._eid)
-        return (this._store.y[this._eid] = n)
+        dirty[this.entity] = true
+        return (this.store.y[this.entity] = n)
       },
       configurable: true
     },
     _z: {
       get() {
-        return this._store.z[this._eid]
+        return this.store.z[this.entity]
       },
       set(n) {
-        dirty.add(this._eid)
-        return (this._store.z[this._eid] = n)
+        dirty[this.entity] = true
+        return (this.store.z[this.entity] = n)
       },
       configurable: true
     },
     _w: {
       get() {
-        return this._store.w[this._eid]
+        return this.store.w[this.entity]
       },
       set(n) {
-        dirty.add(this._eid)
-        return (this._store.w[this._eid] = n)
+        dirty[this.entity] = true
+        return (this.store.w[this.entity] = n)
       },
       configurable: true
     }