--- a/packages/engine/src/ecs/classes/World.ts
+++ b/packages/engine/src/ecs/classes/World.ts
@@ -85,7 +85,6 @@ export class World {
     addComponent(this.cameraEntity, CameraComponent)
     addComponent(this.cameraEntity, VisibleComponent, true)
     setTransformComponent(this.cameraEntity)
-    setLocalTransformComponent(this.cameraEntity, this.originEntity)
 
     /** @todo */
     // this.scene.matrixAutoUpdate = false
@@ -248,7 +247,7 @@ export class World {
     return this.getOwnedNetworkObjectWithComponent(Engine.instance.userId, LocalInputTagComponent) || UndefinedEntity
   }
 
-  dirtyTransforms = new Set<Entity>()
+  readonly dirtyTransforms = {} as Record<Entity, true>
 
   inputState = new Map<InputAlias, InputValue>()
   prevInputState = new Map<InputAlias, InputValue>()