--- a/packages/engine/src/ecs/functions/ComponentFunctions.ts
+++ b/packages/engine/src/ecs/functions/ComponentFunctions.ts
@@ -23,7 +23,7 @@ import { EntityReactorProps, EntityReactorRoot } from './EntityFunctions'
 
 const logger = multiLogger.child({ component: 'engine:ecs:ComponentFunctions' })
 
-const INITIAL_COMPONENT_SIZE = config.client.appEnv === 'test' ? 100000 : 1000 // TODO set to 0 after next bitECS update
+const INITIAL_COMPONENT_SIZE = config.client.appEnv === 'test' ? 100000 : 5000 // TODO set to 0 after next bitECS update
 bitECS.setDefaultSize(INITIAL_COMPONENT_SIZE)
 
 /**
@@ -41,14 +41,18 @@ type StringLiteral<T> = string extends T ? SomeStringLiteral : string
 
 export interface ComponentPartial<
   ComponentType = any,
-  Schema = unknown,
+  Schema extends bitECS.ISchema = {},
   JSON = ComponentType,
   SetJSON = PartialIfObject<DeepReadonly<JSON>>,
   ErrorTypes = never
 > {
   name: string
   schema?: Schema
-  onInit?: (entity: Entity, world: World) => ComponentType & OnInitValidateNotState<ComponentType>
+  onInit?: (
+    this: SoAComponentType<Schema>,
+    entity: Entity,
+    world: World
+  ) => ComponentType & OnInitValidateNotState<ComponentType>
   toJSON?: (entity: Entity, component: State<ComponentType>) => JSON
   onSet?: (entity: Entity, component: State<ComponentType>, json?: SetJSON) => void
   onRemove?: (entity: Entity, component: State<ComponentType>) => void | Promise<void>
@@ -57,14 +61,18 @@ export interface ComponentPartial<
 }
 export interface Component<
   ComponentType = any,
-  Schema = unknown,
+  Schema extends bitECS.ISchema = {},
   JSON = ComponentType,
   SetJSON = PartialIfObject<DeepReadonly<JSON>>,
   ErrorTypes = string
 > {
   name: string
   schema?: Schema
-  onInit: (entity: Entity, world: World) => ComponentType & OnInitValidateNotState<ComponentType>
+  onInit: (
+    this: SoAComponentType<Schema>,
+    entity: Entity,
+    world: World
+  ) => ComponentType & OnInitValidateNotState<ComponentType>
   toJSON: (entity: Entity, component: State<ComponentType>) => JSON
   onSet: (entity: Entity, component: State<ComponentType>, json?: SetJSON) => void
   onRemove: (entity: Entity, component: State<ComponentType>) => void
@@ -127,7 +135,7 @@ export const createMappedComponent = <ComponentType = {}, Schema extends bitECS.
 
 export const getOptionalComponentState = <ComponentType>(
   entity: Entity,
-  component: Component<ComponentType, unknown, unknown>,
+  component: Component<ComponentType, {}, unknown>,
   world = Engine.instance.currentWorld
 ): State<ComponentType> | undefined => {
   // if (entity === UndefinedEntity) return undefined
@@ -137,7 +145,7 @@ export const getOptionalComponentState = <ComponentType>(
 
 export const getComponentState = <ComponentType>(
   entity: Entity,
-  component: Component<ComponentType, unknown, unknown>,
+  component: Component<ComponentType, {}, unknown>,
   world = Engine.instance.currentWorld
 ): State<ComponentType> => {
   const componentState = getOptionalComponentState(entity, component, world)!
@@ -148,7 +156,7 @@ export const getComponentState = <ComponentType>(
 
 export const getOptionalComponent = <ComponentType>(
   entity: Entity,
-  component: Component<ComponentType, unknown, unknown>,
+  component: Component<ComponentType, {}, unknown>,
   world = Engine.instance.currentWorld
 ): ComponentType | undefined => {
   return component.map[entity] as ComponentType | undefined
@@ -156,7 +164,7 @@ export const getOptionalComponent = <ComponentType>(
 
 export const getComponent = <ComponentType>(
   entity: Entity,
-  component: Component<ComponentType, unknown, unknown>,
+  component: Component<ComponentType, {}, unknown>,
   world = Engine.instance.currentWorld
 ): ComponentType => {
   return component.map[entity] as ComponentType