--- a/packages/engine/src/networking/serialization/DataReader.ts
+++ b/packages/engine/src/networking/serialization/DataReader.ts
@@ -185,14 +185,12 @@ export const readBodyRotation = readCompressedRotation(RigidBodyComponent.rotati
 export const readBodyLinearVelocity = readVector3(RigidBodyComponent.linearVelocity)
 export const readBodyAngularVelocity = readVector3(RigidBodyComponent.angularVelocity)
 
-export const readTransform = (v: ViewCursor, entity: Entity, dirtyTransforms: Set<Entity>) => {
+export const readTransform = (v: ViewCursor, entity: Entity, dirtyTransforms: Record<Entity, true>) => {
   const changeMask = readUint8(v)
   let b = 0
   if (checkBitflag(changeMask, 1 << b++)) readPosition(v, entity)
   if (checkBitflag(changeMask, 1 << b++)) readRotation(v, entity)
-  if (entity) {
-    dirtyTransforms.add(entity)
-  }
+  dirtyTransforms[entity] = true
 }
 
 export const readRigidBody = (v: ViewCursor, entity: Entity) => {