--- a/packages/engine/src/physics/components/RigidBodyComponent.ts
+++ b/packages/engine/src/physics/components/RigidBodyComponent.ts
@@ -12,8 +12,6 @@ const QuaternionSchema = { x: f64, y: f64, z: f64, w: f64 }
 const SCHEMA = {
   previousPosition: Vector3Schema,
   previousRotation: QuaternionSchema,
-  previousLinearVelocity: Vector3Schema,
-  previousAngularVelocity: Vector3Schema,
   position: Vector3Schema,
   rotation: QuaternionSchema,
   linearVelocity: Vector3Schema,
@@ -29,8 +27,6 @@ export const RigidBodyComponent = defineComponent({
       body: null! as RigidBody,
       previousPosition: proxifyVector3(this.previousPosition, entity),
       previousRotation: proxifyQuaternion(this.previousRotation, entity),
-      previousLinearVelocity: proxifyVector3(this.previousLinearVelocity, entity),
-      previousAngularVelocity: proxifyVector3(this.previousAngularVelocity, entity),
       position: proxifyVector3(this.position, entity),
       rotation: proxifyQuaternion(this.rotation, entity),
       linearVelocity: proxifyVector3(this.linearVelocity, entity),