… packages/client-core/tsconfig.json
… packages/client/tsconfig.json
… packages/common/tsconfig.json
… packages/editor/tsconfig.json
… packages/engine/src/scene/functions/loaders/ColliderFunctions.ts
… packages/engine/src/transform/systems/TransformSystem.ts
… packages/engine/src/xr/XRAnchorComponents.ts
… packages/engine/src/xr/XRAnchorSystem.ts
… packages/engine/src/xr/XRSessionFunctions.ts
… packages/engine/tsconfig.json
… packages/hyperflux/tsconfig.json
… packages/instanceserver/tsconfig.json
… packages/matchmaking/tsconfig.json
… packages/projects/default-project/tsconfig.json
… packages/projects/template-project/tsconfig.json
… packages/projects/tsconfig.json
… packages/server-core/tsconfig.json
… packages/server/tsconfig.json
… tsconfig.json